requirements.txt
pandas==1.3.4
scikit-learn
scaleogram==0.9.5
scipy==1.10.1
numpy==1.23.5
librosa
seaborn==0.12.2
opencv-python==4.7.0.72
PyWavelets==1.4.1
soundfile==0.12.1
torchsummary==1.5.1
plotly==5.13.1
plotly-express==0.4.1
scikit-image==0.19.3
natsort==8.2.0
h5py
torchviz==0.0.2
prettytable==3.6.0
pyAudioAnalysis
tqdm==4.62.3
moviepy
audiomentations==0.30.0
pydub
torch-audiomentations
PyQt5==5.13.2
opencv-python==4.7.0.72
essentia==2.1b6.dev1034
tables==3.8.0
scikit-learn-intelex==2023.2.1
imblearn
xgboost==1.7.6
lightgbm==4.1.0
itsdangerous==2.0.1
requests==2.31.0
pyarrow
fastparquet
dpctl
hyperopt==0.2.7
imageio==2.4.1
imageio-ffmpeg
nussl
kymatio
scikit-cuda
cupy-cuda116
